+++ R/ParamSet.R
@@ -291,10 +291,10 @@ ParamSet = R6Class("ParamSet",
     #' satisfied and all dependencies are satisfied. Params for which
     #' dependencies are not satisfied should be set to `NA` in `xdt`.
     #'
-    #' @param xdt ([data.table::data.table]).
+    #' @param xdt ([data.table::data.table] | `data.frame()`).
     #' @return If successful `TRUE`, if not a string with the error message.
     check_dt = function(xdt) {
-      xss = transpose(xdt, trafo = FALSE)
+      xss = map(transpose_list(xdt), discard, is.na)
       for (xs in xss) {
         ok = self$check(xs)
         if (!isTRUE(ok)) {
@@ -516,15 +516,15 @@ ParamSet = R6Class("ParamSet",
       private$get_member_with_idnames("is_categ", as.logical)
     },
 
-    #' @field is_numeric (`logical(1)`)\cr
+    #' @field all_numeric (`logical(1)`)\cr
     #' Is `TRUE` if all parameters are [ParamDbl] or [ParamInt].
-    is_numeric = function() {
+    all_numeric = function() {
       all(self$is_number)
     },
 
-    #' @field is_categorical (`logical(1)`)\cr
+    #' @field all_categorical (`logical(1)`)\cr
     #' Is `TRUE` if all parameters are [ParamFct] and [ParamLgl].
-    is_categorical = function() {
+    all_categorical = function() {
       all(self$is_categ)
     },